refactor(devs-infra): shorten import paths
