+++ e2e/__tests__/presets.test.ts
@@ -1,6 +1,17 @@
+import path from 'path'
+
+import execa from 'execa'
+
 import { json as runWithJson } from '../run-jest'
+import { tsJestBundle } from '../utils'
 
 const DIR_NAME = 'presets'
+
+beforeAll(() => {
+  execa.sync('npm', ['install', '--no-package-lock', '--no-shrinkwrap', '--no-save', tsJestBundle], {
+    cwd: path.join(__dirname, '..', DIR_NAME),
+  })
+})
 
 test(`successfully runs the tests inside ${DIR_NAME}`, () => {
   const { json } = runWithJson(DIR_NAME)
