+++ src/cli/cli.spec.ts
@@ -2,7 +2,7 @@ import * as _fs from 'fs'
 import { normalize, resolve } from 'path'
 
 import { logTargetMock, mockObject, mockWriteStream } from '../__helpers__/mocks'
-import { mocked } from '../utils/testing'
+import { mocked } from '../utils'
 
 import { processArgv } from '.'
