+++ src/config/paths-to-module-name-mapper.ts
@@ -2,7 +2,7 @@ import type { Config } from '@jest/types'
 import { LogContexts } from 'bs-logger'
 import type { CompilerOptions } from 'typescript'
 
-import { rootLogger } from '../utils/logger'
+import { rootLogger } from '../utils'
 import { Errors, interpolate } from '../utils/messages'
 
 type TsPathMapping = Exclude<CompilerOptions['paths'], undefined>
