+++ src/ts-jest-transformer.ts
@@ -5,13 +5,11 @@ import type { SyncTransformer, TransformedSource } from '@jest/transform'
 import type { Config } from '@jest/types'
 import type { Logger } from 'bs-logger'
 
-import { TsJestCompiler } from './compiler/ts-jest-compiler'
-import { ConfigSet } from './config/config-set'
+import { TsJestCompiler } from './compiler'
+import { ConfigSet } from './config'
 import { DECLARATION_TYPE_EXT, JS_JSX_REGEX, TS_TSX_REGEX } from './constants'
 import type { CompilerInstance, DepGraphInfo, ProjectConfigTsJest, TransformOptionsTsJest } from './types'
-import { parse, stringify } from './utils/json'
-import { JsonableValue } from './utils/jsonable-value'
-import { rootLogger } from './utils/logger'
+import { parse, stringify, JsonableValue, rootLogger } from './utils'
 import { Errors, interpolate } from './utils/messages'
 import { sha1 } from './utils/sha1'
 import { VersionCheckers } from './utils/version-checkers'
